ReadMe.txt
Written the complete manual test suit for the mate1 website 
Written few automated test cases covering basic process flow of the website(Using Selenium Webdriver in Java). 
Could not convert all the manual test cases into automated as I came accross some trouble while adding external libraries to the
project as few of the libraries I got from the internet were incomplete.
Have mentioned the steps for setting up the complete project and have also attached the Zip file with all the necessary Jars file 
in the GIT repository.
Have hard coded the input values for the text fields in the code. 
The most critical issues in this task were to analyze the iframe pop-up and the mouse over features.
In the final few lines of code I have validated the error of the Text area for "About Myself" section.